requirements.txt
Django==3.2.8
django-crispy-forms==1.13.0
djangorestframework==3.12.4
easy-mask==1.3.4
gunicorn==20.1.0
idna==3.3
jedi==0.18.0
parso==0.8.2
Pillow==8.2.0
ptyprocess==0.7.0
Pygments==2.10.0
python-dotenv==0.19.1
sqlparse==0.4.2
toml==0.10.2
traitlets==4.3.3
urllib3==1.26.7
wcwidth==0.2.5
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
decorator==5.1.0
Django==3.2.8
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
djangorestframework==3.12.4
easy-mask==1.3.4
gunicorn==20.1.0
h11==0.12.0
idna==3.3
jedi==0.18.0
matplotlib-inline==0.1.3
outcome==1.1.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==21.0.0
python-dotenv==0.19.1
